Break up an overly large query-zcl.
